requirements.txt
sqlalchemy~=2.0.41
psycopg2-binary~=2.9.10
boto3~=1.38.43
python-dotenv~=1.1.1
pandas~=2.3.0
pyarrow
